bug 1171957 - Move slugify to mdn.utils
