+++ mdn/tests/test_utils.py
@@ -2,7 +2,7 @@
 """Test mdn.utils."""
 from __future__ import unicode_literals
 
-from mdn.utils import join_content
+from mdn.utils import join_content, slugify
 from .base import TestCase
 
 
@@ -23,3 +23,23 @@ class TestJoinContent(TestCase):
     def test_join_content_empty(self):
         joined = join_content(["Text with", "", "blank", None, "bits", "."])
         self.assertEqual(joined, "Text with blank bits.")
+
+
+class TestSlugify(TestCase):
+    def test_already_slugged(self):
+        self.assertEqual('foo', slugify('foo'))
+
+    def test_long_string(self):
+        self.assertEqual(
+            'abcdefghijklmnopqrstuvwxyz-abcdefghijklmnopqrstuvw',
+            slugify('ABCDEFGHIJKLMNOPQRSTUVWXYZ-abcdefghijklmnopqrstuvwxyz'))
+
+    def test_non_ascii(self):
+        self.assertEqual('_', slugify("Рекомендация"))
+
+    def test_limit(self):
+        self.assertEqual(
+            'abcdefghij', slugify('ABCDEFGHIJKLMNOPQRSTUVWXYZ', length=10))
+
+    def test_num_suffix(self):
+        self.assertEqual('slug13', slugify('slug', suffix=13))
